Add support for specifying a constant_keyword value
For the `constant_keyword` field type Elasticsearch allows for the value to be specified in the mapping.
If the value is not specified in the mapping then the value is set based on the first document that is
indexed.

Package data streams will be able to explicitly declare a constant_keyword value and will not have to
include the field with every document sent.

https://www.elastic.co/guide/en/elasticsearch//reference/7.12/keyword.html#constant-keyword-params