Add validation to check changelog links (#324)
On occasion the PR link in the changelog is not updated with the correct
number. It is left with xxxx or 0 or with no number. The package could get
merged to main with no check to catch it. This PR adds a validation to check
all the changelog entry's links.

Since the link could technically be any issue tracking system, this change
enforces that the base of the path (which is the number) be a valid number
(not a 0) for a github.com URL.