Refactor validator to separate spec loading code (#388)
Refactor validation to split responsibilities between spec loading
and package validation.

A new `specschema` internal package is created to contain the
functionality to load directories spec. All code related to "yamlschema"
has been moved to `yamlschema`.

It is possible now to load a spec without using it for validation, this can
allow us to use the spec for other uses as package loading.