Add settings for data retention (DLM) (#556)
Add a new configuration file for lifecycle configuration with a single setting
for the data retention period.

This is intended to add support for DLM, but is agnostic of the lifecycle
manager used. More settings related to lifecycle management can be added
to the same file in the future.

This is added to integration data streams and input packages.