Add subobjects item to the fields spec
Field definition supports a new `boolenan` item named `subojects`
that allow user to turn off the expansions of field names
containing dots into actual tree of objects.

For example if this is set to `true`, a field named `foo.bar` will
be expanded into an object with a field named `bar` inside an object
named `foo`.