Compliance test suite (#572)
Define an initial compliance suite framework for the Package Spec,
to evaluate if a given stack implements a given version of the spec.

The compliance suite is defined in the `compliance` directory. Tests
are defined in Gherkin syntax (aka Cucumber), and executed using
Godog. Features can be found in the `compliance/features` directory.
Each scenario for each feature is tagged with the package spec version
that it requires to work.

Step definitions can be found in `compliance_test.go`. Additional Go
code is used to initialize Kibana and Elasticsearch clients. There is also
a wrapper for elastic-package, used for build and installation of packages.

Reference packages used on each scenario can be found in
`compliance/testdata/packages.