Require type group or nested for fields with subfields (#629)
Fields with subfields should not have a definition on their own. This would lead
to ambiguity if for example an object would generate a dynamic mapping, but
also definitions for their subelements.

The only cases where a field can contain subfields is when it is of type group,
or nested.