Refactor tests for field definitions (#700)
Adds a new test for field definitions, whose cases receive as inputs a template
package, a set of fields definitions and a list of expected errors. It can also be
customized with an specific spec version or with patches for files.

The test builds for each case a new package using the template package and
following the info in the test case, then it runs the validations and checks for
the expected errors.