Add support for definitions related to deployment modes (#738)
Add support for definitions related to deployment modes. It adds
two main features:
* Policy templates can indicate on what deployment modes they
  can be used.
* Variables can be hidden per deployment mode.

It will be released in two versions:
* 3.1.4 will add support for `agentless` mode.
* 3.2.0 will add settings for `default` mode.

`default` mode is introduced in the next minor because it will need
changes in the stack to be properly supported, while the changes
for agentless will merely be ignored in stacks that don't support it.