Check format of dataset test variable (#753)
`data_stream.dataset` variable has a special meaning in Fleet. It is used to define
the dataset part in the data stream name.
As this is going to end up being part of the data stream name, its format should
be restricted.
Not following this restriction can lead to issues, as the ones described in elastic/kibana#182952.
To avoid these issues, Kibana 8.14.0 will be more restrictive on the formats
allowed (elastic/kibana#182925).
Check the pattern already in the spec to avoid problems later with the validations
in Kibana.