Initial definition of content pack[age]s (#777)
On this initial version, they only support the distribution of kibana assets, like dashboards.

This package type introduces a discovery mechanism based on the data the package can
work with. For example the following definition would indicate that the package provides
resources that are useful when used with data that includes the field `process.pid`:
```
discovery:
  fields:
    - name: "process.pid"
```

The package registry could add support to look for packages that have support for certain
fields. For example a request like the following one could be used to discover packages that
include resources that can leverage the `process.pid` or the `user.id` fields.
```
GET /search?discovery=fields:process.pid,user.id
```