Allow to use group-nested wherever nested can be used
nested and group-nested have different implementations in Fleet, for
nested it doesn't generate any mapping for the subfields, but for
group-nested it does. Integrations are using nested expecting that the
subfields are going to have mappings.

Add group-nested to the spec, as a synonymous of nested, so integrations
can use it when supporting older versions of the stack. nested type
should be fixed in Fleet to work also the same as group-nested.