Add spec for shared folder and link files (#888)
Add a spec definition for a shared folder under `_dev` that will allow users to put
in there any files that are going to be linked from multiple data_streams and/or
packages.

Add a spec definition to allow `*.link` files under pipelines, agent, and fields to let
users reuse repeated files from other parts of the integrations repository.
Links are only allowed when validating from local directory, they are not allowed
in zip packages, that are assumed to be already built.

Adds a test package for both.