Add definition to restrict deployment modes per input (#915)
In various Kibana issues, we found that agentless integrations needed
specific inputs to be allowed/blocked. This is currently in a blocklist
maintained in Fleet, but as more integrations become available for
agentless, it is unsustainable for Fleet to maintain this list. Therefore
the package spec needs to allow integrations to control what inputs are
allowed on agentless, and Fleet should read from this information.