feat(spec): Add duration variable data type
Adds a new duration variable data type that allows defining time-based
values like "60s", "1m", or "2h45m". The valid time units are "ms", "s", "m",
and "h".

This change introduces two new validation properties for this type:
- min_duration: The minimum allowed duration.
- max_duration: The maximum allowed duration.

The default value for a duration variable must also conform to the
duration format.

The 'd' (day) unit is intentionally not supported. This design decision
mirrors the Go language's time.Duration type, which avoids units of
a day or larger. The ambiguity of a "day" (which can be 23, 24, or 25
hours long due to daylight saving time) can lead to subtle bugs in
time-based calculations. By omitting it, we enforce more precise and
less ambiguous time interval definitions.

Relates #44