Reserve otelcol input type (#975)
We are using "otelcol" input type as convention for input
packages that configure OTel collector components. This
requires support in fleet-server and Fleet, that will be
released in 9.2.

By reserving this input type we are preventing its use in
GA packages for older stacks.

It will be allowed for input packages starting on package
spec 3.5.0.