Input packages don't require to define fields
With OTel inputs we have seen that all fields required are already
defined in the built-in templates, and requiring to add mappings can
lead to conflicts with them.

For ECS inputs, required fields are also defined in ecs@mappings, so
there would be no need to require them.

These packages can still define fields if needed.

Test package for OTel is updated for current recommendations.