Allow packages which have format_version < 3.5.0 to ship alert rule templates.
This allows Integrations to ship templates without bumping format_version explicitly.
The impact of this is that packages can still support installation on 8.19.x (where
templates are ignored), whilst also providing templates for users on stack versions
that support them.

It leaves the decision about format_version up to Integration developers, which ends
up being a trade off between stack support for the package and clarity for users about
what features and functionality the package provides.