Fix rule matching processor for event.original handling
In #1013 the spec for the processors required for event.original
handling was modified to allow to remove the message field when included
in an array. This modified the error message, that doesn't match anymore
with the rule that allows to skip this test.

Update the rules to match with this new pattern of error messages.