Fix rule for processors handing event.original so the check can be skipped (#1024)
In #1013 the spec for the processors required for event.original
handling was modified to allow to remove the message field when included
in an array. This modified the error message by adding the index in the
array, what doesn't match anymore with the rule that allows to skip this
test.

Update the rules to match with this new pattern of error messages.

Another rule is added to allow to skip this check in pipelines that only
have a remove processor, that doesn't have if or ignore_missing
properties.