Include the validation of transform fields (#1027)
Include the validation of the fields defined in transforms as
part of the rest of the fields (data streams and input packages).
Thereby, fields in transforms will run the same validations as the
other fields with one exception about the required fields. It is not
required to defined `data_stream.{type,namespace,dataset}` and 
`@timestamp` fields in transforms.