Handlebars template static validation and documentation (#1030)
Includes a validation rule for verifying the static .hbs files used at input tempate and stream template are correct.
Also adds more information on the spec regarding the format of these templates and the helpers that are available at fleet when rendering them.
Linked file are also verified .hbs.link

The validation is done over the static syntax of handlebars, does not validate the render of the template taking into account the values of the values.